Split multitile AP code
