+++ applications/stlp/src/app_conf.h
@@ -28,6 +28,9 @@
 /* If in channel sample format, appconfAUDIO_PIPELINE_FRAME_ADVANCE == MIC_ARRAY_CONFIG_SAMPLES_PER_FRAME*/
 #define appconfAUDIO_PIPELINE_FRAME_ADVANCE     MIC_ARRAY_CONFIG_SAMPLES_PER_FRAME
 
+#define appconfINPUT_SAMPLES_MIC_DELAY_MS 150
+
+
 #ifndef appconfI2S_ENABLED
 #if XVF3610_Q60A || XCOREAI_EXPLORER
 #define appconfI2S_ENABLED         1
