+++ applications/stlp/src/audio_pipeline/aec/aec_memory_pool.h
@@ -3,11 +3,7 @@
 #ifndef AEC_MEMORY_POOL_H
 #define AEC_MEMORY_POOL_H
 
-#define AEC_MAX_Y_CHANNELS        (2)
-#define AEC_MAX_X_CHANNELS        (2)
-#define AEC_MAIN_FILTER_PHASES    (10)
-#define AEC_SHADOW_FILTER_PHASES  (5)
-
+#include "audio_pipeline/aec/aec_config.h"
 #include "xs3_math.h"
 
 //Memory pool definition
