--- applications/stlp/src/audio_pipeline/audio_pipeline.c
@@ -1,356 +1,0 @@
-// Copyright (c) 2022 XMOS LIMITED. This Software is subject to the terms of the
-// XMOS Public License: Version 1
-
-/* STD headers */
-#include <string.h>
-#include <stdint.h>
-#include <xcore/hwtimer.h>
-
-/* FreeRTOS headers */
-#include "FreeRTOS.h"
-#include "task.h"
-#include "timers.h"
-#include "queue.h"
-#include "stream_buffer.h"
-
-/* Library headers */
-#include "generic_pipeline.h"
-#include "aec_api.h"
-#include "agc_api.h"
-#include "ic_api.h"
-#include "ns_api.h"
-#include "vad_api.h"
-#include "audio_pipeline/aec_memory_pool.h"
-#include "adec_api.h"
-
-/* App headers */
-#include "app_conf.h"
-#include "app_control/app_control.h"
-#include "audio_pipeline/audio_pipeline.h"
-
-extern void aec_process_frame_1thread(
-        aec_state_t *main_state,
-        aec_state_t *shadow_state,
-        int32_t (*output_main)[AEC_FRAME_ADVANCE],
-        int32_t (*output_shadow)[AEC_FRAME_ADVANCE],
-        const int32_t (*y_data)[AEC_FRAME_ADVANCE],
-        const int32_t (*x_data)[AEC_FRAME_ADVANCE]);
-
-#define appconfINPUT_SAMPLES_MIC_DELAY_MS 150
-
-/**
- * Delay buffer contains the configured buffer delay size + 1 frame worth of audio
- * When appconfINPUT_SAMPLES_MIC_DELAY_MS  is positive, the mic is delayed
- * When appconfINPUT_SAMPLES_MIC_DELAY_MS is negative, the mic is "advanced" by delaying ref
- */
-#define ABS(A) ((A >= 0) ? A : -A)
-#define AP_INPUT_SAMPLES_MIC_DELAY_SIZE_PER_CHAN        ( 16000*ABS(appconfINPUT_SAMPLES_MIC_DELAY_MS)/1000 )
-#define AP_INPUT_SAMPLES_MIC_DELAY_CHAN_CNT             ( (appconfINPUT_SAMPLES_MIC_DELAY_MS > 0) ? AEC_MAX_Y_CHANNELS : AEC_MAX_X_CHANNELS )
-#define AP_INPUT_SAMPLES_MIC_DELAY_SIZE_CHAN            ( AP_INPUT_SAMPLES_MIC_DELAY_SIZE_PER_CHAN * AP_INPUT_SAMPLES_MIC_DELAY_CHAN_CNT )
-#define AP_INPUT_SAMPLES_MIC_DELAY_SIZE_CUR_FRAME_WORDS ( AP_INPUT_SAMPLES_MIC_DELAY_CHAN_CNT * appconfAUDIO_PIPELINE_FRAME_ADVANCE)
-#define AP_INPUT_SAMPLES_MIC_DELAY_CUR_FRAME_BYTES      ( AP_INPUT_SAMPLES_MIC_DELAY_SIZE_CUR_FRAME_WORDS * sizeof(int32_t))
-#define AP_INPUT_SAMPLES_MIC_DELAY_BUF_SIZE_BYTES       ( AP_INPUT_SAMPLES_MIC_DELAY_SIZE_CHAN * sizeof(int32_t) )
-
-/* Note: Changing the order here will effect the channel order for
- * audio_pipeline_input() and audio_pipeline_output()
- */
-typedef struct {
-    int32_t samples[appconfAUDIO_PIPELINE_CHANNELS][appconfAUDIO_PIPELINE_FRAME_ADVANCE];
-    int32_t aec_reference_audio_samples[appconfAUDIO_PIPELINE_CHANNELS][appconfAUDIO_PIPELINE_FRAME_ADVANCE];
-    int32_t mic_samples_passthrough[appconfAUDIO_PIPELINE_CHANNELS][appconfAUDIO_PIPELINE_FRAME_ADVANCE];
-
-    /* Below is additional context needed by other stages on a per frame basis */
-    uint8_t vad;
-    float_s32_t max_ref_energy;
-    float_s32_t aec_corr_factor;
-} frame_data_t;
-
-#if appconfAUDIO_PIPELINE_FRAME_ADVANCE != 240
-#error This pipeline is only configured for 240 frame advance
-#endif
-
-typedef struct stage_delay_ctx {
-    StreamBufferHandle_t delay_buf;
-} stage_delay_ctx_t;
-
-typedef struct aec_ctx {
-    aec_state_t DWORD_ALIGNED aec_main_state;
-    aec_state_t DWORD_ALIGNED aec_shadow_state;
-    aec_shared_state_t DWORD_ALIGNED aec_shared_state;
-    uint8_t DWORD_ALIGNED aec_main_memory_pool[sizeof(aec_memory_pool_t)];
-    uint8_t DWORD_ALIGNED aec_shadow_memory_pool[sizeof(aec_shadow_filt_memory_pool_t)];
-} aec_ctx_t;
-
-typedef struct ic_stage_ctx {
-    ic_state_t DWORD_ALIGNED state;
-} ic_stage_ctx_t;
-
-typedef struct vad_stage_ctx {
-    vad_state_t DWORD_ALIGNED state;
-} vad_stage_ctx_t;
-
-typedef struct ns_stage_ctx {
-    ns_state_t DWORD_ALIGNED state;
-} ns_stage_ctx_t;
-
-typedef struct agc_stage_ctx {
-    agc_meta_data_t md;
-    agc_state_t DWORD_ALIGNED state;
-} agc_stage_ctx_t;
-
-#if ON_TILE(1)
-static stage_delay_ctx_t delay_buf_state = {};
-static aec_ctx_t aec_state = {};
-#endif
-
-static ic_stage_ctx_t ic_stage_state = {};
-static vad_stage_ctx_t vad_stage_state = {};
-static ns_stage_ctx_t ns_stage_state = {};
-static agc_stage_ctx_t agc_stage_state = {};
-
-static void *audio_pipeline_input_i(void *input_app_data)
-{
-    frame_data_t *frame_data;
-
-    frame_data = pvPortMalloc(sizeof(frame_data_t));
-
-#if ON_TILE(0)
-    size_t bytes_received = 0;
-    bytes_received = rtos_intertile_rx_len(
-            intertile_ctx,
-            appconfAUDIOPIPELINE_PORT,
-            portMAX_DELAY);
-
-    xassert(bytes_received == sizeof(frame_data_t));
-
-    rtos_intertile_rx_data(
-            intertile_ctx,
-            frame_data,
-            bytes_received);
-#endif
-
-#if ON_TILE(1)
-    audio_pipeline_input(input_app_data,
-                       (int32_t **)frame_data->aec_reference_audio_samples,
-                       4,
-                       appconfAUDIO_PIPELINE_FRAME_ADVANCE);
-
-    frame_data->vad = 0;
-
-    memcpy(frame_data->samples, frame_data->mic_samples_passthrough, sizeof(frame_data->samples));
-#endif
-    return frame_data;
-}
-
-static int audio_pipeline_output_i(frame_data_t *frame_data,
-                                   void *output_app_data)
-{
-
-#if ON_TILE(1)
-    rtos_intertile_tx(intertile_ctx,
-                      appconfAUDIOPIPELINE_PORT,
-                      frame_data,
-                      sizeof(frame_data_t));
-    return AUDIO_PIPELINE_FREE_FRAME;
-#endif
-#if ON_TILE(0)
-    return audio_pipeline_output(output_app_data,
-                               (int32_t **)frame_data->samples,
-                               6,
-                               appconfAUDIO_PIPELINE_FRAME_ADVANCE);
-#endif
-}
-
-#if ON_TILE(1)
-static void stage_delay(frame_data_t *frame_data)
-{
-#if (appconfINPUT_SAMPLES_MIC_DELAY_MS > 0) /* Delay mics */
-    size_t bytes_sent = xStreamBufferSend(
-                                delay_buf_state.delay_buf,
-                                &frame_data->samples,
-                                AP_INPUT_SAMPLES_MIC_DELAY_CUR_FRAME_BYTES,
-                                0);
-
-    configASSERT(bytes_sent == AP_INPUT_SAMPLES_MIC_DELAY_CUR_FRAME_BYTES);
-
-    if (xStreamBufferBytesAvailable(delay_buf_state.delay_buf) >= AP_INPUT_SAMPLES_MIC_DELAY_BUF_SIZE_BYTES) {
-        size_t bytes_rx = xStreamBufferReceive(
-                                    delay_buf_state.delay_buf,
-                                    &frame_data->samples,
-                                    AP_INPUT_SAMPLES_MIC_DELAY_CUR_FRAME_BYTES,
-                                    0);
-
-        configASSERT(bytes_rx == AP_INPUT_SAMPLES_MIC_DELAY_CUR_FRAME_BYTES);
-    }
-#elif (appconfINPUT_SAMPLES_MIC_DELAY_MS < 0)/* Delay Ref*/
-    size_t bytes_sent = xStreamBufferSend(
-                                delay_buf_state.delay_buf,
-                                &frame_data->aec_reference_audio_samples,
-                                AP_INPUT_SAMPLES_MIC_DELAY_CUR_FRAME_BYTES,
-                                0);
-
-    configASSERT(bytes_sent == AP_INPUT_SAMPLES_MIC_DELAY_CUR_FRAME_BYTES);
-
-    if (xStreamBufferBytesAvailable(delay_buf_state.delay_buf) >= AP_INPUT_SAMPLES_MIC_DELAY_BUF_SIZE_BYTES) {
-        size_t bytes_rx = xStreamBufferReceive(
-                                    delay_buf_state.delay_buf,
-                                    &frame_data->aec_reference_audio_samples,
-                                    AP_INPUT_SAMPLES_MIC_DELAY_CUR_FRAME_BYTES,
-                                    0);
-
-        configASSERT(bytes_rx == AP_INPUT_SAMPLES_MIC_DELAY_CUR_FRAME_BYTES);
-    }
-#else  /* Do nothing */
-#endif
-}
-
-static void stage_1(frame_data_t *frame_data)
-{
-    int32_t stage1_output[AEC_MAX_Y_CHANNELS][appconfAUDIO_PIPELINE_FRAME_ADVANCE];
-    aec_process_frame_1thread(
-            &aec_state.aec_main_state,
-            &aec_state.aec_shadow_state,
-            stage1_output,
-            NULL,
-            frame_data->samples,
-            frame_data->aec_reference_audio_samples);
-
-    frame_data->max_ref_energy = aec_calc_max_input_energy(
-                                    frame_data->aec_reference_audio_samples,
-                                    aec_state.aec_main_state.shared_state->num_x_channels);
-    frame_data->aec_corr_factor = aec_calc_corr_factor(aec_state.aec_main_state, 0);
-    memcpy(frame_data->samples, stage1_output, AEC_MAX_Y_CHANNELS * appconfAUDIO_PIPELINE_FRAME_ADVANCE * sizeof(int32_t));
-}
-#endif
-
-static void stage_vad_and_ic(frame_data_t *frame_data)
-{
-    int32_t ic_output[appconfAUDIO_PIPELINE_FRAME_ADVANCE];
-
-    ic_filter(&ic_stage_state.state,
-              frame_data->samples[0],
-              frame_data->samples[1],
-              ic_output);
-    uint8_t vad = vad_probability_voice(ic_output, &vad_stage_state.state);
-    ic_adapt(&ic_stage_state.state, vad, ic_output);
-    frame_data->vad = vad;
-    memcpy(frame_data->samples, ic_output, appconfAUDIO_PIPELINE_FRAME_ADVANCE * sizeof(int32_t));
-}
-
-static void stage_ns(frame_data_t *frame_data)
-{
-    int32_t ns_output[appconfAUDIO_PIPELINE_FRAME_ADVANCE];
-    configASSERT(NS_FRAME_ADVANCE == appconfAUDIO_PIPELINE_FRAME_ADVANCE);
-    ns_process_frame(
-                &ns_stage_state.state,
-                ns_output,
-                frame_data->samples[0]);
-    memcpy(frame_data->samples, ns_output, appconfAUDIO_PIPELINE_FRAME_ADVANCE * sizeof(int32_t));
-}
-
-static void stage_agc(frame_data_t *frame_data)
-{
-    int32_t agc_output[appconfAUDIO_PIPELINE_FRAME_ADVANCE];
-    configASSERT(AGC_FRAME_ADVANCE == appconfAUDIO_PIPELINE_FRAME_ADVANCE);
-
-    agc_stage_state.md.vad_flag = (frame_data->vad > AGC_VAD_THRESHOLD);
-    agc_stage_state.md.aec_ref_power = frame_data->max_ref_energy;
-    agc_stage_state.md.aec_corr_factor = frame_data->aec_corr_factor;
-
-    agc_process_frame(
-            &agc_stage_state.state,
-            agc_output,
-            frame_data->samples[0],
-            &agc_stage_state.md);
-    memcpy(frame_data->samples, agc_output, appconfAUDIO_PIPELINE_FRAME_ADVANCE * sizeof(int32_t));
-}
-
-static void initialize_pipeline_stages(void)
-{
-#if ON_TILE(1)
-    configASSERT(AP_INPUT_SAMPLES_MIC_DELAY_BUF_SIZE_BYTES > 0);
-
-    delay_buf_state.delay_buf = xStreamBufferCreate((size_t)AP_INPUT_SAMPLES_MIC_DELAY_BUF_SIZE_BYTES + AP_INPUT_SAMPLES_MIC_DELAY_CUR_FRAME_BYTES, 0);
-    configASSERT(delay_buf_state.delay_buf);
-
-    aec_init(&aec_state.aec_main_state,
-             &aec_state.aec_shadow_state,
-             &aec_state.aec_shared_state,
-             &aec_state.aec_main_memory_pool[0],
-             &aec_state.aec_shadow_memory_pool[0],
-             AEC_MAX_Y_CHANNELS,
-             AEC_MAX_X_CHANNELS,
-             AEC_MAIN_FILTER_PHASES,
-             AEC_SHADOW_FILTER_PHASES);
-#endif
-#if ON_TILE(0)
-    ic_init(&ic_stage_state.state);
-    vad_init(&vad_stage_state.state);
-    ns_init(&ns_stage_state.state);
-    agc_init(&agc_stage_state.state, &AGC_PROFILE_ASR);
-    agc_stage_state.md.vad_flag = AGC_META_DATA_NO_VAD;
-    agc_stage_state.md.aec_ref_power = AGC_META_DATA_NO_AEC;
-    agc_stage_state.md.aec_corr_factor = AGC_META_DATA_NO_AEC;
-#endif
-}
-
-void audio_pipeline_init(
-    void *input_app_data,
-    void *output_app_data)
-{
-#if ON_TILE(1)
-    const int stage_count = 2;
-
-    const pipeline_stage_t stages[] = {
-        (pipeline_stage_t)stage_delay,
-        (pipeline_stage_t)stage_1,
-    };
-
-    const configSTACK_DEPTH_TYPE stage_stack_sizes[] = {
-        configMINIMAL_STACK_SIZE + RTOS_THREAD_STACK_SIZE(stage_delay) + RTOS_THREAD_STACK_SIZE(audio_pipeline_input_i),
-        configMINIMAL_STACK_SIZE + RTOS_THREAD_STACK_SIZE(stage_1) + RTOS_THREAD_STACK_SIZE(audio_pipeline_output_i),
-    };
-
-    initialize_pipeline_stages();
-
-    app_control_aec_servicer_register();
-
-    generic_pipeline_init((pipeline_input_t)audio_pipeline_input_i,
-                        (pipeline_output_t)audio_pipeline_output_i,
-                        input_app_data,
-                        output_app_data,
-                        stages,
-                        (const size_t*) stage_stack_sizes,
-                        appconfAUDIO_PIPELINE_TASK_PRIORITY,
-                        stage_count);
-#endif
-#if ON_TILE(0)
-    const int stage_count = 3;
-
-    const pipeline_stage_t stages[] = {
-        (pipeline_stage_t)stage_vad_and_ic,
-        (pipeline_stage_t)stage_ns,
-        (pipeline_stage_t)stage_agc,
-    };
-
-    const configSTACK_DEPTH_TYPE stage_stack_sizes[] = {
-        configMINIMAL_STACK_SIZE + RTOS_THREAD_STACK_SIZE(stage_vad_and_ic) + RTOS_THREAD_STACK_SIZE(audio_pipeline_input_i),
-        configMINIMAL_STACK_SIZE + RTOS_THREAD_STACK_SIZE(stage_ns),
-        configMINIMAL_STACK_SIZE + RTOS_THREAD_STACK_SIZE(stage_agc) + RTOS_THREAD_STACK_SIZE(audio_pipeline_output_i),
-    };
-
-    initialize_pipeline_stages();
-
-    app_control_stage1_servicer_register();
-    app_control_stage2_servicer_register();
-
-    generic_pipeline_init((pipeline_input_t)audio_pipeline_input_i,
-                        (pipeline_output_t)audio_pipeline_output_i,
-                        input_app_data,
-                        output_app_data,
-                        stages,
-                        (const size_t*) stage_stack_sizes,
-                        appconfAUDIO_PIPELINE_TASK_PRIORITY,
-                        stage_count);
-#endif
-}
